requirements.txt
dico-api>=0.0.32
dico-command
dico-interaction
dico-dp
aiomysql